+++ pytorch_hccl_tests/osu/collectives/osu_allgather.py
@@ -21,7 +21,7 @@ def allgather(args):
     rank = dist.get_rank()
     world_size = dist.get_world_size()
     dtype = args.dtype
-    device = get_device(backend, rank)
+    device = get_device(backend, args.local_rank)
     pg = None
 
     options = Options("Allgather", args)