+++ pytorch_hccl_tests/osu/collectives/osu_alltoall.py
@@ -22,7 +22,7 @@ def alltoall(args):
     rank = dist.get_rank()
     world_size = dist.get_world_size()
     dtype = args.dtype
-    device = get_device(backend, rank)
+    device = get_device(backend, args.local_rank)
 
     options = Options("Alltoall", args)
     Utils.check_numprocs(world_size, rank, limit=3)