+++ pytorch_hccl_tests/osu/collectives/osu_barrier.py
@@ -14,7 +14,7 @@ def barrier(args):
     backend = args.backend
     rank = dist.get_rank()
     world_size = dist.get_world_size()
-    device = get_device(backend, rank)
+    device = get_device(backend, args.local_rank)
 
     options = Options("Barrier", args)
     Utils.check_numprocs(world_size, rank, limit=3)